+++ server/server.js
@@ -1,14 +1,9 @@
 const express = require('express')
 const app = express()
 const port = 3000
-var mysql = require('mysql');
+var connection = require('./config');
 
-var connection = mysql.createConnection({
-  host     : 'localhost',
-  user     : 'root',
-  password : 'beapineapple',
-  database : ''
-});
+
  
 app.get('/rodney', (req, res) => {
     connection.query('SELECT * FROM ___ ', function (error, results) {
